refactor: export httpServer from app.ts for testing
